Relevant BUGIDs:
Purpose of commit: cleanup

Commit summary:
---------------

2008-01-28  Thorsten Kukuk  <kukuk@thkukuk.de>

        * libpam/pam_audit.c: Include pam_modutil_private.h.

        * libpam/pam_item.c (pam_set_item): Fix compiler warning.

        * libpam/pam_end.c (pam_end): Cast to correct pointer type.
        * libpam/include/security/_pam_macros.h (_pam_overwrite_n): Use
        unsigned int.

+++ libpam/pam_item.c
@@ -148,7 +148,8 @@ int pam_set_item (pam_handle_t *pamh, int item_type, const void *item)
 	    free(pamh->xauth.name);
 	}
 	if (pamh->xauth.datalen) {
-	    _pam_overwrite_n(pamh->xauth.data, pamh->xauth.datalen);
+	  _pam_overwrite_n(pamh->xauth.data,
+			   (unsigned int) pamh->xauth.datalen);
 	    free(pamh->xauth.data);
 	}
 	pamh->xauth = *((const struct pam_xauth_data *) item);
